requirements.txt
django==3.2.12
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2-binary==2.8.6
gunicorn==20.1.0
django-heroku==0.3.1